Move entirely to async
Instead of having each connection monitored by a thread, spawn a task
that waits on the connecion's reader and writer task. The main loop then
waits on this top-level task and uses it as a signal to reap the connection.

This avoids a blocking Join on every connection close, and also avoids
reaping connections on a timer since those connections can be awaited.